requirements.txt
Flask==3.0.0
requests==2.27.1
requests-toolbelt==0.9.1
python-dotenv==0.20.0
openai==1.73.0
pyyaml==6.0.2
apscheduler==3.11.0